requirements.txt
yfinance==0.2.31
pandas==2.1.4
python-telegram-bot==20.3
ta==0.11.0
httpx==0.25.2
python-dotenv==1.0.0
pytz==2023.3.post1
psutil==5.9.6
tzdata==2023.3  # For timezone support
